Support resources that live at FTP URLs.
Handle it if GET() returns an empty status code, which happens with FTP
URLs (see https://github.com/hadley/httr/issues/190).
Also handle it if the status code is 302, which just means there was a 
redirect.

This still produces a lot of output from the FTP server. 
Also if Rsamtools is not installed, downloading a dbSNP VCF resource
will cause an error, so some sort of dependency on Rsamtools needs to be declared.