Bug fix: ISNAN not supported with integers.
An error occurs at build time with gcc-6.1 because ISNAN takes an int
as argument. This operation is not supported, but interestingly this
error is not reported by gcc-5.3 or lower. The solution is to test equality
of the int with NA_INTEGER (which is implemented as INT_MIN in R).